requirements.txt
# AutoMecanik RAG Service - Dependencies
# Version 100% gratuite (sans API payantes)

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Embeddings (GRATUIT - local, FastEmbed = Rust backend, ~80% moins de RAM)
fastembed>=0.2.0

# Vector Database
weaviate-client>=4.4.0

# Cache
redis>=5.0.0

# HTTP Client
httpx>=0.26.0

# LLM (Claude API - V2 spec)
anthropic>=0.18.0

# LangGraph (Pipeline orchestration - V2 spec)
langgraph>=0.0.40
langchain-core>=0.1.0

# Markdown parsing
python-frontmatter>=1.1.0
pypdf>=4.0.0

# Utils
numpy>=1.24.0

# YAML Config (Quarantine Mode)
pyyaml>=6.0

# DB connectors (ingestion DB views)
psycopg[binary]>=3.2.0
PyMySQL>=1.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Admin Interface
jinja2>=3.1.0
python-multipart>=0.0.9
aiofiles>=23.0.0
